Add "#" before keyword automatically
Plurk's searching API ignores "#" when searching, and that's the reason
we have to check if `keyword` is in `content_raw" specifically, because
the search result will also include plurks have the same words but
doesn't have the hashtag.

Now the script will automatically add "#" to the given keyword, and it
sanitizes the parameter first, so we will not have redundant "#" either.